+++ src/components/CardExpirationDate/CardExpirationDate.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { MONTH, LIMIT_LENGTH, LEADING_ZERO_MONTHS } from '../../../../constants';
-import { inputNumberOnly, limitInputLength } from '../../../../utils';
+import { MONTH, LIMIT_LENGTH, LEADING_ZERO_MONTHS } from '../../constants';
+import { inputNumberOnly, limitInputLength } from '../../utils';
 
 function CardExpirationDate({ cardInfo, setCardInfo }) {
   const handleMonthOnFocusOut = (event) => {
