+++ src/components/CardInputForm/CardInputForm.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import checkCardInfo from '../../../../validator';
+import checkCardInfo from '../../validator';
 
 function CardInputForm({ cardInfo, children }) {
   const handleOnSubmit = (e) => {
