+++ src/components/CardOwner/CardOwner.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { LIMIT_LENGTH } from '../../../../constants';
-import { limitInputLength, inputEnglishOnly } from '../../../../utils';
+import { LIMIT_LENGTH } from '../../constants';
+import { limitInputLength, inputEnglishOnly } from '../../utils';
 
 function CardOwner({ cardInfo, setCardInfo }) {
   const handleOnChange = (event) => {
