+++ src/components/CardPassword/CardPassword.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { inputNumberOnly, limitInputLength } from '../../../../utils';
-import { LIMIT_LENGTH } from '../../../../constants';
+import { inputNumberOnly, limitInputLength } from '../../utils';
+import { LIMIT_LENGTH } from '../../constants';
 
 function CardPassword({ cardInfo, setCardInfo }) {
   const handleOnChange = (event) => {
