+++ src/components/Tooltip/Tooltip.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { CVC_DESCRIPTION } from '../../../../constants';
+import { CVC_DESCRIPTION } from '../../constants';
 
 function Tooltip() {
   return (
